# Slotted enumeration faster