--- a/src/Paprika.Tests/Data/SlottedArrayTests.cs
+++ b/src/Paprika.Tests/Data/SlottedArrayTests.cs
@@ -60,6 +60,7 @@ public class SlottedArrayTests
         map.GetAssert(key0, Data0);
         map.GetAssert(key1, Data1);
         map.GetAssert(key2, Data2);
+        map.GetAssert(key3, Data3);
         map.GetAssert(key4, Data4);
 
         using var e = map.EnumerateAll();
@@ -83,6 +84,61 @@ public class SlottedArrayTests
         e.MoveNext().Should().BeTrue();
         e.Current.Key.Equals(key4).Should().BeTrue();
         e.Current.RawData.SequenceEqual(Data4).Should().BeTrue();
+
+        e.MoveNext().Should().BeFalse();
+
+        // verify
+        return Verify(span.ToArray());
+    }
+
+    [Test]
+    public Task Enumerate_nibble([Values(1, 2, 3, 4)] int nibble)
+    {
+        Span<byte> span = stackalloc byte[256];
+        var map = new SlottedArray(span);
+
+        var key0 = NibblePath.Empty;
+        var key1 = NibblePath.FromKey(stackalloc byte[1] { 0x1A });
+        var key2 = NibblePath.FromKey(stackalloc byte[2] { 0x2A, 13 });
+        var key3 = NibblePath.FromKey(stackalloc byte[3] { 0x3A, 13, 31 });
+        var key4 = NibblePath.FromKey(stackalloc byte[4] { 0x4A, 13, 31, 41 });
+
+        map.SetAssert(key0, Data0);
+        map.SetAssert(key1, Data1);
+        map.SetAssert(key2, Data2);
+        map.SetAssert(key3, Data3);
+        map.SetAssert(key4, Data4);
+
+        map.GetAssert(key0, Data0);
+        map.GetAssert(key1, Data1);
+        map.GetAssert(key2, Data2);
+        map.GetAssert(key3, Data3);
+        map.GetAssert(key4, Data4);
+
+        var expected = nibble switch
+        {
+            1 => key1,
+            2 => key2,
+            3 => key3,
+            4 => key4,
+            _ => throw new Exception()
+        };
+
+        var data = nibble switch
+        {
+            1 => Data1,
+            2 => Data2,
+            3 => Data3,
+            4 => Data4,
+            _ => throw new Exception()
+        };
+
+        using var e = map.EnumerateNibble((byte)nibble);
+
+        e.MoveNext().Should().BeTrue();
+
+        e.Current.Key.Equals(expected).Should().BeTrue();
+        e.Current.RawData.SequenceEqual(data).Should().BeTrue();
 
         e.MoveNext().Should().BeFalse();