… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_nibble_nibble=1.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_nibble_nibble=2.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_nibble_nibble=3.verified.bin
… src/Paprika.Tests/Data/SlottedArrayTests.Enumerate_nibble_nibble=4.verified.bin