This PR introduces two changes to enumeration over `SlottedArray`:

1. `EnumerateAll` that enumerates now with a bit less overhead (code size reduction and some inlines)
2. `EnumerateNibble` that uses the new encoding of length to to leverage the filtering and returns data without having their keys unprepared if not needed. It's not 16 faster, but gives ~3x performance boost for scans.

### Benchmarks

For `EnumerateAll` the existing one is used. For `EnumerateNibble` new benchmarks were added as previously it was filtering over the materialized key which was not different from `EnumerateAll`.

#### Before

| Method       | Mean     | Error   | StdDev  | Code Size |
|------------- |---------:|--------:|--------:|----------:|
| EnumerateAll | 751.6 ns | 7.62 ns | 6.75 ns |   1,913 B |

#### After

| Method          | nibble | Mean     | Error   | StdDev  | Code Size |
|---------------- |------- |---------:|--------:|--------:|----------:|
| EnumerateAll    | ?      | 744.7 ns | 8.57 ns | 8.01 ns |   1,883 B |
| EnumerateNibble | 0      | 239.3 ns | 2.72 ns | 2.54 ns |   1,939 B |
| EnumerateNibble | 1      | 241.3 ns | 4.42 ns | 4.14 ns |   1,932 B |